security: derive identity for the knowledge-base query and the rag-id write too
Sentry flagged askKnowledgeBase on this PR as a CRITICAL IDOR, and it is
right: askKnowledgeBase(userId, query, projectUuid) let any caller aim
the retrieval at another tenant's documents and read back an answer
composed from them. updateDocRagId(docUuid, ragId, userId) is the same
shape on the write side — any caller could repoint another tenant's rag
document.

Both have non-browser callers that establish identity themselves: the MCP
connector from an API key, and app/api/documents/ai from its own auth. So
they take the same split as the reads — implementations in
lib/library/queries.ts keeping an explicit userId, session-deriving
wrappers in app/actions/library.ts taking none.

I had deferred this one as its own PR. Having it reported against the
very file this PR is about made shipping without it the wrong call.

The #210 baseline drops from 59 to 53: all six library.ts entries, which
is the whole of what this PR closes stated from the outside.
