+++ tests/security/library-session-identity.test.ts
@@ -7,14 +7,21 @@ vi.mock('@/lib/library/queries', () => ({
   getDocByUuidFor: vi.fn(async () => null),
   getDocumentVersionsFor: vi.fn(async () => ({ success: true, versions: [] })),
   getProjectStorageUsageFor: vi.fn(async () => ({ success: true, fileStorage: 0 })),
+  askKnowledgeBaseFor: vi.fn(async () => ({ success: true, answer: 'x' })),
+  updateDocRagIdFor: vi.fn(async () => ({ success: true })),
   WORKSPACE_STORAGE_LIMIT: 100 * 1024 * 1024,
 }));
 
 const { getServerSession } = vi.mocked(await import('next-auth'));
 const queries = vi.mocked(await import('@/lib/library/queries'));
-const { getDocs, getDocByUuid, getDocumentVersions, getProjectStorageUsage } = await import(
-  '@/app/actions/library'
-);
+const {
+  getDocs,
+  getDocByUuid,
+  getDocumentVersions,
+  getProjectStorageUsage,
+  askKnowledgeBase,
+  updateDocRagId,
+} = await import('@/app/actions/library');
 
 const VICTIM = 'victim-user-id';
 const CALLER = 'caller-user-id';
@@ -101,5 +108,46 @@ describe('library actions derive identity from the session', () => {
     await expect(getProjectStorageUsage()).resolves.toMatchObject({ success: false });
 
     expect(queries.getDocsFor).not.toHaveBeenCalled();
+  });
+
+  it('runs the knowledge-base query as the session user', async () => {
+    // askKnowledgeBase(userId, …) let any caller aim the retrieval at another
+    // tenant's documents and read the answer built from them.
+    await askKnowledgeBase('what is in my docs', 'project-uuid');
+
+    expect(queries.askKnowledgeBaseFor).toHaveBeenCalledWith(
+      CALLER,
+      'what is in my docs',
+      'project-uuid'
+    );
+    expect(queries.askKnowledgeBaseFor).not.toHaveBeenCalledWith(
+      VICTIM,
+      expect.anything(),
+      expect.anything()
+    );
+  });
+
+  it('refuses the knowledge-base query without a session', async () => {
+    getServerSession.mockResolvedValue(null as any);
+
+    const result = await askKnowledgeBase('anything');
+
+    expect(result.success).toBe(false);
+    expect(queries.askKnowledgeBaseFor).not.toHaveBeenCalled();
+  });
+
+  it('writes the rag id as the session user', async () => {
+    await updateDocRagId('doc-uuid', 'rag-id');
+
+    expect(queries.updateDocRagIdFor).toHaveBeenCalledWith(CALLER, 'doc-uuid', 'rag-id');
+  });
+
+  it('refuses the rag id write without a session', async () => {
+    getServerSession.mockResolvedValue(null as any);
+
+    const result = await updateDocRagId('doc-uuid', 'rag-id');
+
+    expect(result.success).toBe(false);
+    expect(queries.updateDocRagIdFor).not.toHaveBeenCalled();
   });
 });
