+++ tests/security/unguarded-actions.json
@@ -14,18 +14,12 @@
   "app/actions/custom-mcp-servers.ts#toggleCustomMcpServerStatus",
   "app/actions/custom-mcp-servers.ts#updateCustomMcpServer",
   "app/actions/discover-mcp-tools.ts#discoverSingleServerToolsInternal",
-  "app/actions/library.ts#askKnowledgeBase",
   "app/actions/library.ts#createDoc",
   "app/actions/library.ts#deleteDoc",
-  "app/actions/library.ts#getDocByUuid",
-  "app/actions/library.ts#getDocs",
-  "app/actions/library.ts#getDocumentVersions",
-  "app/actions/library.ts#getProjectStorageUsage",
   "app/actions/library.ts#manualRepairDocumentRagIds",
   "app/actions/library.ts#reindexDocument",
   "app/actions/library.ts#repairMissingRagDocumentIds",
   "app/actions/library.ts#trackDocumentView",
-  "app/actions/library.ts#updateDocRagId",
   "app/actions/log-retention.ts#cleanupMcpServerLogs",
   "app/actions/log-retention.ts#updateLogRetentionPolicy",
   "app/actions/mcp-playground.ts#addServerLogForProfile",
